+++ .idea/gradle.xml
@@ -6,6 +6,7 @@
       <GradleProjectSettings>
         <option name="testRunner" value="CHOOSE_PER_TEST" />
         <option name="externalProjectPath" value="$PROJECT_DIR$/AccountBook_JAVA_edit_ver" />
+        <option name="gradleHome" value="$USER_HOME$/gradle/gradle-8.4" />
         <option name="gradleJvm" value="#GRADLE_LOCAL_JAVA_HOME" />
         <option name="modules">
           <set>
