+++ AccountBook_JAVA_edit_ver/app/build.gradle.kts
@@ -51,6 +51,7 @@ dependencies {
 
     implementation("io.github.ParkSangGwon:tedpermission-normal:3.3.0")
     implementation("androidx.exifinterface:exifinterface:1.3.7")
+    implementation ("org.jetbrains.kotlinx:kotlinx-coroutines-android:1.5.2")
