+++ AccountBook_JAVA_edit_ver/app/src/main/AndroidManifest.xml
@@ -20,6 +20,7 @@
 
     <application
         android:networkSecurityConfig="@xml/network_security_config"
+        android:usesCleartextTraffic="true"
         android:allowBackup="true"
         android:dataExtractionRules="@xml/data_extraction_rules"
         android:fullBackupContent="@xml/backup_rules"
