+++ AccountBook_JAVA_edit_ver/app/src/main/java/com/example/accountbook_java_edit_ver/ServerIP.java
@@ -1,5 +1,5 @@
 package com.example.accountbook_java_edit_ver;
 
 public class ServerIP {
-    public static final String SERVER_IP = "10.0.2.2";
+    public static final String SERVER_IP = "http://192.168.133.137:8080/";
 }
